[GH-390] Fix #390 incorrect escaping of initial_root_password
* Return raw password from `root_password` instead of shell-escaping
* SQL-escape the password variable before rendering in `init_records_script`
* Generate the mysql init records script as a quoted heredoc

The previous code was broken for some passwords because it used shell
escaping for all values, based on historical behaviour that sent the passwords
to mysql on the command line.

It was not updated to take account of the difference between shell and SQL
escaping when the cookbook was changed to generate a SQL file and pipe it
to mysql.

Mysql usually ignores \ escapes in string literals - `\x` becomes `x`, `\&`
becomes `&`. However some - notably `%` - are shell-escaped to `\%` but
**stay as `\%` when parsed by mysql**.

See http://dev.mysql.com/doc/refman/5.7/en/string-literals.html

eg: `PASSWORD('pa\$\$w%d')` === `pa$$w%d`

The previous implementation was assigning an incorrect root password on
initialisation, making the instance unusable.

This fixes the issue by only escaping explicit `\` and `'` characters in
the password string and rendering in a quoted heredoc so that shell
escaping is not required.

Note: this is not a perfect SQL quoting solution - it may break with nulls,
tabs and other non-printable special characters, but it seems reasonable to
assume these are unlikely to be used in passwords.